Fix test as modules moved
